backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
supabase>=2.10.0
httpx<0.28.0
bcrypt>=4.1.2
requests>=2.32.0
pandas>=2.2.0
reportlab>=4.0.9
python-dateutil>=2.8.2
holidays>=0.40
qrcode[pil]>=7.4.2
pillow>=10.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
anthropic>=0.25.0
cryptography
pytest
